Split promotion tests to sequential and parallel
- add globalSetup script to set app promotion admin
- move calculatePalleteAddress helper to DevUniqueHelper
